# mock-server/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Clinic OS API Specification (SCS)
  description: |
    المستند البرمجي لواجهات برمجة التطبيقات (API Specification) لمنصة عيادتي الذكية.
    يحتوي الملف على مسارين أساسيين:
    1. **واجهات الأوبريشن والمشغل (Platform Operations):** تبدأ بـ `/admin/v1` ومحمية بـ VPN.
    2. **واجهات العيادات والعملاء (Clinic Dashboard):** تبدأ بـ `/v1` وتعتمد على ترويسة العيادة.
  version: 1.0.0
servers:
  - url: https://api.SCS-admin.com
    description: Public Clinic Dashboard Gateway
  - url: https://api.SCS-ops.com
    description: Private Operations Gateway (Accessible via VPN only)
  - url: http://localhost:3000
    description: Local Mock Development Server
paths:
...
  /webhooks/whatsapp:
    get:
      summary: التحقق من رابط الـ Webhook الخاص بواتساب (Meta Verification)
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: تأكيد نجاح التحقق وإرجاع الـ challenge
          content:
            text/plain:
              schema:
                type: string
                example: "12345678"

    post:
      summary: استقبال رسائل الواتساب من المرضى (Meta WhatsApp Webhook Receiver)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: تم الاستلام والمعالجة
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
  /webhooks/telegram:
    post:
      summary: استقبال رسائل تليجرام من المرضى (Telegram Bot Webhook Receiver)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: تم الاستلام
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true